+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Add a new `TiledColliderCreated` event and an associated observer, when a new collider is spawned.
 
+### Bugfixes
+
+- Do not restrict `world_chunking` system to camera with `Changed<Transform>` (#94)
+
 ## v0.7.0
 
 **BREAKING CHANGES**